requirements.txt
albumentations==1.3.0
arabic_reshaper==2.1.4
coremltools==6.1
Flask==2.2.2
ipython==8.6.0
matplotlib==3.6.2
numpy==1.23.5
onnx==1.12.0
onnxruntime==1.13.1
onnxsim==0.4.10
opencv_python==4.6.0.66
pafy==0.5.5
pandas==1.5.1
Pillow==9.3.0
plotly==5.11.0
protobuf==4.21.9
psutil==5.9.0
pycocotools==2.0.6
python_bidi==0.4.2
PyYAML==6.0
requests==2.28.1
scipy==1.9.3
seaborn==0.12.1
setuptools==65.5.0
streamlit==1.15.0
tensorflow==2.11.0
tensorflowjs==4.1.0
tensorrt==8.5.1.7
tflite_runtime==2.10.0
thop==0.1.1.post2209072238
torch==1.13.0
torchvision==0.14.0
tqdm==4.64.1
